chore(datetime): Rename component and paths #1841
