+++ src/app/scheduling/datetimeeditor/datetime-advanced/datetime-advanced.html
@@ -1,4 +1,4 @@
-<div class="date-time-wrapper">
+<div class="datetime-wrapper">
     <label>Increment/Decrement Default</label>
     <igx-input-group>
         <input igxInput #dateEditorDefault="igxDateTimeEditor" type="text" [igxDateTimeEditor]="'dd/MM/yyyy'" [(ngModel)]="date">
