+++ src/app/scheduling/datetimeeditor/datetime-advanced/datetime-advanced.scss
@@ -1,4 +1,4 @@
-.date-time-wrapper {
+.datetime-wrapper {
     padding: 16px;
     width: 300px;
